% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMul.R
\name{matmatvecMul}
\alias{matmatvecMul}
\title{Title}
\usage{
matmatvecMul(
  x,
  y,
  z,
  mode = 1,
  x_row = dim(x)[1],
  x_col = dim(x)[2],
  y_row = dim(y)[1],
  y_col = dim(y)[2],
  z_dim = length(z)
)
}
\arguments{
\item{z_dim}{}
}
\description{
Title
}
