% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMul.R
\name{matMulmat}
\alias{matMulmat}
\title{Title}
\usage{
matMulmat(
  x,
  y,
  x_row = dim(x)[1],
  x_col = dim(x)[2],
  y_row = dim(y)[1],
  y_col = dim(y)[2]
)
}
\arguments{
\item{y_col}{}
}
\description{
Title
}
